Fix #944: Show alert when trying to remove device while offline.
