+++ Client/Frontend/Sync/SyncViewController.swift
@@ -29,21 +29,11 @@ class SyncViewController: UIViewController {
         NotificationCenter.default.removeObserver(self)
     }
     
-    func showNoConnectionAlert() {
-        let alert = UIAlertController(title: Strings.SyncNoConnectionTitle,
-                                      message: Strings.SyncNoConnectionBody, preferredStyle: .alert)
-        
-        let okAction = UIAlertAction(title: Strings.OKString, style: .default)
-        alert.addAction(okAction)
-        
-        present(alert, animated: true)
-    }
-    
     /// Perform a block of code only if user has a network connection, shows an error alert otherwise.
     /// Most of sync initialization methods require an internet connection.
     func doIfConnected(code: () -> Void) {
         if !DeviceInfo.hasConnectivity() {
-            showNoConnectionAlert()
+            present(SyncAlerts.noConnection, animated: true)
             return
         }
