+++ Client/Frontend/Sync/SyncWelcomeViewController.swift
@@ -140,11 +140,7 @@ class SyncWelcomeViewController: SyncViewController {
             
             let bvc = (UIApplication.shared.delegate as? AppDelegate)?.browserViewController
             
-            let title = Strings.SyncInitErrorTitle
-            let message = Strings.SyncInitErrorMessage
-            let alert = UIAlertController(title: title, message: message, preferredStyle: .alert)
-            alert.addAction(UIAlertAction(title: Strings.OKString, style: .default, handler: nil))
-            bvc?.present(alert, animated: true)
+            bvc?.present(SyncAlerts.initializationError, animated: true)
         }
     }
